Validate input is a supergraph in `Supergraph::new`
