+++ src/lib.rs
@@ -1,12 +1,5 @@
 #![allow(dead_code)] // TODO: This is fine while we're iterating, but should be removed later.
 
-use crate::error::FederationError;
-use crate::merge::merge_subgraphs;
-use crate::merge::MergeFailure;
-use crate::subgraph::ValidSubgraph;
-use apollo_compiler::validation::Valid;
-use apollo_compiler::Schema;
-
 mod api_schema;
 mod compat;
 pub mod error;
@@ -17,10 +10,53 @@ pub mod query_plan;
 pub mod schema;
 pub mod subgraph;
 
-pub use api_schema::ApiSchemaOptions;
-pub use query_graph::extract_subgraphs_from_supergraph::ValidFederationSubgraph;
-pub use query_graph::extract_subgraphs_from_supergraph::ValidFederationSubgraphs;
-use schema::ValidFederationSchema;
+pub use crate::api_schema::ApiSchemaOptions;
+use crate::error::FederationError;
+use crate::error::SingleFederationError;
+use crate::link::join_spec_definition::JoinSpecDefinition;
+use crate::link::link_spec_definition::LinkSpecDefinition;
+use crate::link::spec::Identity;
+use crate::merge::merge_subgraphs;
+use crate::merge::MergeFailure;
+pub use crate::query_graph::extract_subgraphs_from_supergraph::ValidFederationSubgraph;
+pub use crate::query_graph::extract_subgraphs_from_supergraph::ValidFederationSubgraphs;
+use crate::schema::ValidFederationSchema;
+use crate::subgraph::ValidSubgraph;
+use apollo_compiler::validation::Valid;
+use apollo_compiler::Schema;
+use link::join_spec_definition::JOIN_VERSIONS;
+use schema::FederationSchema;
+
+pub(crate) type SupergraphSpecs = (&'static LinkSpecDefinition, &'static JoinSpecDefinition);
+
+/// Checks that required supergraph directives are in the schema, and returns which ones were used.
+pub(crate) fn validate_supergraph(
+    supergraph_schema: &FederationSchema,
+) -> Result<SupergraphSpecs, FederationError> {
+    let Some(metadata) = supergraph_schema.metadata() else {
+        return Err(SingleFederationError::InvalidFederationSupergraph {
+            message: "Invalid supergraph: must be a core schema".to_owned(),
+        }
+        .into());
+    };
+    let link_spec_definition = metadata.link_spec_definition()?;
+    let Some(join_link) = metadata.for_identity(&Identity::join_identity()) else {
+        return Err(SingleFederationError::InvalidFederationSupergraph {
+            message: "Invalid supergraph: must use the join spec".to_owned(),
+        }
+        .into());
+    };
+    let Some(join_spec_definition) = JOIN_VERSIONS.find(&join_link.url.version) else {
+        return Err(SingleFederationError::InvalidFederationSupergraph {
+            message: format!(
+                "Invalid supergraph: uses unsupported join spec version {} (supported versions: {})",
+                JOIN_VERSIONS.versions().map(|v| v.to_string()).collect::<Vec<_>>().join(", "),
+                join_link.url.version,
+            ),
+        }.into());
+    };
+    Ok((link_spec_definition, join_spec_definition))
+}
 
 pub struct Supergraph {
     pub schema: ValidFederationSchema,
@@ -29,13 +65,18 @@ pub struct Supergraph {
 impl Supergraph {
     pub fn new(schema_str: &str) -> Result<Self, FederationError> {
         let schema = Schema::parse_and_validate(schema_str, "schema.graphql")?;
-        // TODO: federation-specific validation
         Self::from_schema(schema)
     }
 
     pub fn from_schema(schema: Valid<Schema>) -> Result<Self, FederationError> {
+        let schema = schema.into_inner();
+        let schema = FederationSchema::new(schema)?;
+
+        let _ = validate_supergraph(&schema)?;
+
         Ok(Self {
-            schema: ValidFederationSchema::new(schema)?,
+            // We know it's valid because the input was.
+            schema: schema.assume_valid(),
         })
     }
 
@@ -70,7 +111,6 @@ mod tests {
 
     #[test]
     fn can_extract_subgraph() {
-        // TODO: not actually implemented; just here to give a sense of the API.
         let schema = r#"
           schema
             @link(url: "https://specs.apollo.dev/link/v1.0")
@@ -151,7 +191,8 @@ mod tests {
         "#;
 
         let supergraph = Supergraph::new(schema).unwrap();
-        let _subgraphs = crate::query_graph::extract_subgraphs_from_supergraph::extract_subgraphs_from_supergraph(&supergraph.schema, None)
+        let _subgraphs = supergraph
+            .extract_subgraphs()
             .expect("Should have been able to extract subgraphs");
         // TODO: actual assertions on the subgraph once it's actually implemented.
     }
